chore: separate + fix type definitions
