+++ lib/array.d.ts
@@ -1,0 +1,8 @@
+/**
+ * Flatten array, one level deep.
+ *
+ * @param arr
+ *
+ * @return
+ */
+export function flatten<T extends any[]>(arr: T[]): T;
