+++ lib/array.js
@@ -2,9 +2,11 @@
 /**
  * Flatten array, one level deep.
  *
- * @param {Array<?>} arr
+ * @template T
  *
- * @return {Array<?>}
+ * @param {T[][]} arr
+ *
+ * @return {T[]}
  */
 export function flatten(arr) {
   return Array.prototype.concat.apply([], arr);
