+++ lib/collection.js
@@ -6,23 +6,74 @@ import {
   has
 } from './lang.js';
 
+/**
+ * @template T
+ * @typedef { (
+ *   ((e: T) => boolean) |
+ *   ((e: T, idx: number) => boolean) |
+ *   ((e: T, key: string) => boolean) |
+ *   string |
+ *   number
+ * ) } Matcher
+ */
+
+/**
+ * @template T
+ * @template U
+ *
+ * @typedef { (
+ *   ((e: T) => U) | string | number
+ * ) } Extractor
+ */
+
+
+/**
+ * @template T
+ * @typedef { (val: T, key: any) => boolean } MatchFn
+ */
+
+/**
+ * @template T
+ * @typedef { T[] } ArrayCollection
+ */
+
+/**
+ * @template T
+ * @typedef { { [key: string]: T } } StringKeyValueCollection
+ */
+
+/**
+ * @template T
+ * @typedef { { [key: number]: T } } NumberKeyValueCollection
+ */
+
+/**
+ * @template T
+ * @typedef { StringKeyValueCollection<T> | NumberKeyValueCollection<T> } KeyValueCollection
+ */
+
+/**
+ * @template T
+ * @typedef { KeyValueCollection<T> | ArrayCollection<T> } Collection
+ */
 
 /**
  * Find element in collection.
  *
- * @param  {Array|Object} collection
- * @param  {Function|Object} matcher
+ * @template T
+ * @param {Collection<T>} collection
+ * @param {Matcher<T>} matcher
  *
  * @return {Object}
  */
 export function find(collection, matcher) {
 
-  matcher = toMatcher(matcher);
+  const matchFn = toMatcher(matcher);
 
   let match;
 
   forEach(collection, function(val, key) {
-    if (matcher(val, key)) {
+    if (matchFn(val, key)) {
       match = val;
 
       return false;
@@ -37,19 +88,20 @@ export function find(collection, matcher) {
 /**
  * Find element index in collection.
  *
- * @param  {Array|Object} collection
- * @param  {Function} matcher
+ * @template T
+ * @param {Collection<T>} collection
+ * @param {Matcher<T>} matcher
  *
- * @return {Object}
+ * @return {number}
  */
 export function findIndex(collection, matcher) {
 
-  matcher = toMatcher(matcher);
+  const matchFn = toMatcher(matcher);
 
   let idx = isArray(collection) ? -1 : undefined;
 
   forEach(collection, function(val, key) {
-    if (matcher(val, key)) {
+    if (matchFn(val, key)) {
       idx = key;
 
       return false;
@@ -61,19 +113,22 @@ export function findIndex(collection, matcher) {
 
 
 /**
- * Find element in collection.
+ * Filter elements in collection.
  *
- * @param  {Array|Object} collection
- * @param  {Function} matcher
+ * @template T
+ * @param {Collection<T>} collection
+ * @param {Matcher<T>} matcher
  *
- * @return {Array} result
+ * @return {T[]} result
  */
 export function filter(collection, matcher) {
+
+  const matchFn = toMatcher(matcher);
 
   let result = [];
 
   forEach(collection, function(val, key) {
-    if (matcher(val, key)) {
+    if (matchFn(val, key)) {
       result.push(val);
     }
   });
@@ -86,10 +141,11 @@ export function filter(collection, matcher) {
  * Iterate over collection; returning something
  * (non-undefined) will stop iteration.
  *
- * @param  {Array|Object} collection
- * @param  {Function} iterator
+ * @template T
+ * @param {Collection<T>} collection
+ * @param { ((item: T, idx: number) => (boolean|void)) | ((item: T, key: string) => (boolean|void)) } iterator
  *
- * @return {Object} return result that stopped the iteration
+ * @return {T} return result that stopped the iteration
  */
 export function forEach(collection, iterator) {
 
@@ -119,10 +175,11 @@ export function forEach(collection, iterator) {
 /**
  * Return collection without element.
  *
- * @param  {Array} arr
- * @param  {Function} matcher
+ * @template T
+ * @param {ArrayCollection<T>} arr
+ * @param {Matcher<T>} matcher
  *
- * @return {Array}
+ * @return {T[]}
  */
 export function without(arr, matcher) {
 
@@ -132,10 +189,10 @@ export function without(arr, matcher) {
 
   ensureArray(arr);
 
-  matcher = toMatcher(matcher);
+  const matchFn = toMatcher(matcher);
 
   return arr.filter(function(el, idx) {
-    return !matcher(el, idx);
+    return !matchFn(el, idx);
   });
 
 }
@@ -144,11 +201,14 @@ export function without(arr, matcher) {
 /**
  * Reduce collection, returning a single result.
  *
- * @param  {Object|Array} collection
- * @param  {Function} iterator
- * @param  {any} result
+ * @template T
+ * @template V
  *
- * @return {any} result returned from last iterator
+ * @param {Collection<T>} collection
+ * @param {(result: V, entry: T, index: any) => V} iterator
+ * @param {V} result
+ *
+ * @return {V} result returned from last iterator
  */
 export function reduce(collection, iterator, result) {
 
@@ -252,8 +312,8 @@ export function values(collection) {
 /**
  * Group collection members by attribute.
  *
- * @param  {Object|Array} collection
- * @param  {Function} extractor
+ * @param {Object|Array} collection
+ * @param {Extractor} extractor
  *
  * @return {Object} map with { attrValue => [ a, b, c ] }
  */
@@ -300,8 +360,10 @@ export const unionBy = uniqueBy;
 /**
  * Sort collection by criteria.
  *
- * @param  {Object|Array} collection
- * @param  {String|Function} extractor
+ * @template T
+ *
+ * @param {Collection<T>} collection
+ * @param {Extractor<T, number | string>} extractor
  *
  * @return {Array}
  */
@@ -341,13 +403,17 @@ export function sortBy(collection, extractor) {
  *
  * @example
  *
+ * ```javascript
  * const matcher = matchPattern({ id: 1 });
  *
  * let element = find(elements, matcher);
+ * ```
  *
- * @param  {Object} pattern
+ * @template T
  *
- * @return {Function} matcherFn
+ * @param {T} pattern
+ *
+ * @return { (el: any) =>  boolean } matcherFn
  */
 export function matchPattern(pattern) {
 
@@ -361,13 +427,30 @@ export function matchPattern(pattern) {
 }
 
 
+/**
+ * @param {string | ((e: any) => any) } extractor
+ *
+ * @return { (e: any) => any }
+ */
 function toExtractor(extractor) {
+
+  /**
+   * @satisfies { (e: any) => any }
+   */
   return isFunction(extractor) ? extractor : (e) => {
+
+    // @ts-ignore: just works
     return e[extractor];
   };
 }
 
 
+/**
+ * @template T
+ * @param {Matcher<T>} matcher
+ *
+ * @return {MatchFn<T>}
+ */
 function toMatcher(matcher) {
   return isFunction(matcher) ? matcher : (e) => {
     return e === matcher;
