+++ lib/lang.js
@@ -25,6 +25,11 @@ export function isNumber(obj) {
   return nativeToString.call(obj) === '[object Number]';
 }
 
+/**
+ * @param {any} obj
+ *
+ * @return {boolean}
+ */
 export function isFunction(obj) {
   const tag = nativeToString.call(obj);
