+++ lib/object.js
@@ -25,9 +25,13 @@ export function assign(target, ...others) {
  *
  * This mutates the object and returns it.
  *
- * @param {Object} target The target of the set operation.
+ * @template T
+ *
+ * @param {T} target The target of the set operation.
  * @param {(string|number)[]} path The path to the nested value.
  * @param {any} value The value to set.
+ *
+ * @return {T}
  */
 export function set(target, path, value) {
 
@@ -75,6 +79,8 @@ export function set(target, path, value) {
  * @param {Object} target The target of the get operation.
  * @param {(string|number)[]} path The path to the nested value.
  * @param {any} [defaultValue] The value to return if no value exists.
+ *
+ * @return {any}
  */
 export function get(target, path, defaultValue) {
 
@@ -96,12 +102,15 @@ export function get(target, path, defaultValue) {
 }
 
 /**
- * Pick given properties from the target object.
+ * Pick properties from the given target.
  *
- * @param {Object} target
- * @param {Array} properties
+ * @template T
+ * @template {any[]} V
  *
- * @return {Object} target
+ * @param {T} target
+ * @param {V} properties
+ *
+ * @return Pick<T, V>
  */
 export function pick(target, properties) {
 
@@ -122,10 +131,13 @@ export function pick(target, properties) {
 /**
  * Pick all target properties, excluding the given ones.
  *
- * @param {Object} target
- * @param {Array} properties
+ * @template T
+ * @template {any[]} V
  *
- * @return {Object} target
+ * @param {T} target
+ * @param {V} properties
+ *
+ * @return {Omit<T, V>} target
  */
 export function omit(target, properties) {
